+++ app/utils/colors.js
@@ -4,7 +4,9 @@ const COLORS = {
   WHITE: '#fff',
   GRAY: '#6A6A6A',
   ACCENT: '#143240',
-  SECENDARY: '#41C8F4',
+  SECENDARY: '#4747E6',
+  // SECENDARY: '#41C8F4',
+
   GRADIENT:
     'linear-gradient(221deg, rgba(87,92,203,1) 0%, rgba(76,144,223,1) 95%, rgba(73,162,230,1) 97%, rgba(73,164,231,1) 98%, rgba(70,177,235,1) 99%, rgba(65,200,244,1) 100%)',
 };
